# Backyard-app/openapi.yaml
openapi: 3.0.1
info:
  title: Backyard
  description: This is a Backyard server.
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: pietrusjakub@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'https://backyard.swagger.io/v1'
  - url: 'http://backyard.swagger.io/v1'
...
paths:
...
  /court:
    get:
      tags:
        - court
      summary: Get courts list
      operationId: getCourtsList
      parameters:
        - name: limit
          in: query
          description: The maximum number of results to retrieve
          schema:
            type: integer
        - name: offset
          in: query
          description: The zero-ary offset into the results
          schema:
            type: integer
        - name: sort
          in: query
          description: How to sort the results
          schema:
            type: string
            default: id
            enum:
              - id
              - name
              - location
              - rating
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Court'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Court'
    put:
      tags:
        - court
      summary: Update an existing court
      operationId: updateCourt
      requestBody:
        description: Court object that needs to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Court'
          application/xml:
            schema:
              $ref: '#/components/schemas/Court'
        required: true
      responses:
        '400':
          description: Invalid ID supplied
          content: {}
        '404':
          description: Court not found
          content: {}
        '405':
          description: Validation exception
          content: {}
      x-codegen-request-body-name: body
    post:
      tags:
        - court
      summary: Create court
      operationId: createCourt
      requestBody:
        description: Created court object
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Court'
        required: true
      responses:
        default:
          description: successful operation
          content: {}
      x-codegen-request-body-name: body
  '/court/{courtId}':
    get:
      tags:
        - court
      summary: Finds court by courtId
      operationId: getCourtById
      parameters:
        - name: courtId
          in: path
          description: Id of court to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Court'
            application/xml:
              schema:
                $ref: '#/components/schemas/Court'
        '400':
          description: Invalid ID supplied
          content: {}
        '404':
          description: Court not found
          content: {}
    delete:
      tags:
        - court
      summary: Deletes a court
      operationId: deleteCourt
      parameters:
        - name: courtId
          in: path
          description: Court id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid ID supplied
          content: {}
        '404':
          description: Court not found
          content: {}
  '/court/{courtId}/rate':
    post:
      tags:
        - court
      summary: Rate court
      operationId: rateCourt
      parameters:
        - name: courtId
          in: path
          description: Id of court to rate
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Rate object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourtRating'
          application/xml:
            schema:
              $ref: '#/components/schemas/CourtRating'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourtRating'
            application/xml:
              schema:
                $ref: '#/components/schemas/CourtRating'
        '405':
          description: Invalid input
          content: {}
      x-codegen-request-body-name: body
  '/court/{courtId}/uploadImage':
    post:
      tags:
        - court
      summary: Uploads an image
      operationId: uploadCourtImage
      parameters:
        - name: courtId
          in: path
          description: ID of court to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                additionalMetadata:
                  type: string
                  description: Additional data to pass to server
                file:
                  type: string
                  description: file to upload
                  format: binary
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
...
    Court:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        location:
          type: string
        description:
          type: string
      xml:
        name: Court
...
    CourtRating:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        value:
          type: integer
          format: int32
        surfaceQuality:
          type: integer
          format: int32
        basketsQuality:
          type: integer
          format: int32
        boardsQuality:
          type: integer
          format: int32
      xml:
        name: CourtRating
...